+++ History.md
@@ -1,5 +1,6 @@
 ## In Git
 
+* [Java] GluePath can be a package name ([#164](https://github.com/cucumber/cucumber-jvm/issues/164) Aslak Hellesøy)
 * [Build] Fixed subtle path issues on Windows
 * [Build] Fixed Build Failure: Cucumber-JVM: Scala (FAILURE) ([#167](https://github.com/cucumber/cucumber-jvm/issues/167) Aslak Hellesøy)